+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Changed
 
 * API: `font_` prefix changed to `fcft_`
+* API: renamed `struct glyph` to `struct fcft_glyph`
 * API: `enum subpixel_order` renamed to `enum fcft_subpixel`, and
   `ORDER` was removed from the enum values.